+++ errpref/eprefquark.go
@@ -526,8 +526,9 @@ func (ePrefQuark *errPrefQuark) resetErrPrefDisplayLineLengthToDefault() {
 //       maximum number of characters allowed in a text display
 //       line for error prefix information.
 //
-//       If 'maxErrPrefixTextLineLength' is set to a value of zero
-//       (0), this method will take no action and return.
+//       If 'maxErrPrefixTextLineLength' is set to a value less
+//       than the minimum Error Prefix Line Length, this method
+//       will take no action and return.
 //
 //
 // -----------------------------------------------------------------
@@ -556,7 +557,9 @@ func (ePrefQuark *errPrefQuark) setErrPrefDisplayLineLength(
 
 	defer defaultErrPrefLineLenLock.Unlock()
 
-	if maxErrPrefixStringLength == 0 {
+	minimum := errPrefPreon{}.ptr().getMinErrPrefLineLength()
+
+	if maxErrPrefixStringLength < minimum {
 		return
 	}