+++ errpref/releasenotes.md
@@ -4,7 +4,7 @@ This version of ***errpref*** was compiled and tested using ***Go*** 1.16.3.
 
 This version supports ***Go*** modules.
 
-## Version 1.6.2
+## Version 1.7.0
 
 ##### Documentation Updates
 1. Modified Best Practices documentation for implementing ErrPrefixDto in calls to private, internal methods.
@@ -32,7 +32,30 @@ This version supports ***Go*** modules.
 4. Added new method ErrPrefixDto.ReplaceLastErrPrefix()
    This method deletes and replaces the Last Error Prefix Information object in the current ErrPrefixDto collection 
    with new error prefix and context information.
-   
+
+##### New Leading and Trailing Text Feature
+ErrPrefixDto has added a new feature which allows the user to configure leading and trailing strings which will be
+added to the beginning and ending of error prefix text displays. The leading and trailing text strings may be
+comprised of any string of characters including new lines ('\n'), tabs ('\t') and line separators.
+
+For more information see the documentation on the following methods:
+1. ErrPrefixDto.SetLeadingTextStr()
+2. ErrPrefixDto.SetTrailingTextStr()
+3. ErrPrefixDto.GetLeadingTextStr()
+4. ErrPrefixDto.GetTrailingTextStr()
+5. ErrPrefixDto.ClearLeadingTextStr()
+6. ErrPrefixDto.ClearTrailingTextStr()
+
+##### Enforcing Minimum Error Prefix Line Length
+Users have always had the ability to set the maximum line length for error prefix text displays. As of this release,
+the minimum line length will be enforced by methods in the ErrPrefixDto and ErrPref types. The Minimum Error Prefix
+Line Length is '10' characters. Depending on the function called, any attempt to set a value less than the Minimum
+Error Prefix Line Length will cause a reset to the default Line Length value, or generate an error. For more
+information see the documentation on the following methods:
+1. ErrPref.SetMaxErrPrefTextLineLength()
+2. ErrPrefixDto.SetMaxTextLineLen()
+3. ErrPrefixDto.GetMaxTextLineLen()
+
 ## Version 1.6.1
 
 #### Changes